requirements.txt
# LibrarianRobot_Vision requirements
# Example: pip install -r requirements.txt

## 1. Ultralytics requirements

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.22.2 # pinned by Snyk to avoid a vulnerability
# opencv-python>=4.6.0  # avoid redundancy install
# pillow>=7.1.2         # avoid redundancy install
# pyyaml>=5.3.1         # avoid redundancy install
# requests>=2.23.0
# scipy>=1.4.1
torch>=1.8.0
torchvision>=0.9.0
# tqdm>=4.64.0

# Logging -------------------------------------
# tensorboard>=2.13.0
# dvclive>=2.12.0
# clearml
# comet

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Extras --------------------------------------
psutil  # system utilization
py-cpuinfo  # display CPU info

## 2. parseq requirements

aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.0.4
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
joblib==1.3.2
lightning-utilities==0.9.0
multidict==6.0.4
nltk==3.8.1
packaging==23.1
pillow==10.0.0
pytorch-lightning==1.9.5
pyyaml==6.0.1
regex==2023.8.8
# requests==2.31.0
safetensors==0.3.2
timm==0.9.5
tqdm==4.66.1
typing-extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2

## 3. fast_demo requirements
addict==2.4.0
# certifi==2022.12.7
# charset-normalizer==3.1.0
cmake==3.26.0
Cython==0.29.33
editdistance==0.6.2
# filelock==3.10.0
# idna==3.4
Jinja2==3.1.2
lit==15.0.7
MarkupSafe==2.1.2
mmcv==1.7.1
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.6.0.66
# packaging==23.0
Pillow==9.4.0
Polygon3==3.0.9.1
prefetch-generator==1.0.3
pyclipper==1.3.0.post4
PyYAML==6.0
requests>=2.28.2
scipy==1.10.1
sympy==1.11.1
# tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
# urllib3==1.26.15
yacs==0.1.8
yapf==0.32.0

## other installations
easyocr==1.7.0
pyrealsense2==2.54.1.5217

### mmcv install
# pip install -U openmim
# mim install mmcv